[Add] Added ebay api support
Added ebay dev support and started to get it working within
the desired framework. Currently will take a list of keywords
and price range and then log the returned items in a dictionary
based on the item id. logs the url, id and title of returned items
so that emails can be constructed based upon a comparison with the
results of previous searches saved as a pickles